refactor: 멘션 삭제 시 null 판별 로직 수정

+++ src/main/java/shop/flowchat/notification/command/service/MessageReadModelCommandService.java
@@ -45,8 +45,8 @@ public class MessageReadModelCommandService {
     }
 
     public void deleteMessage(MentionEventPayload payload) {
+        if (!messageRepository.existsById(payload.messageId())) return;
         messageRepository.deleteById(payload.messageId());
         attachmentRepository.deleteByMessageId(payload.messageId());
     }
-
 }
